# docs/openapi.json
{
  "components": {
    "schemas": {
      "AltAzEntry": {
        "additionalProperties": false,
        "description": "Authoritative Skyfield values for one target at one instant.",
        "properties": {
          "alt_deg": {
            "description": "Altitude, degrees (refracted only when requested).",
            "title": "Alt Deg",
            "type": "number"
          },
          "az_deg": {
            "description": "Azimuth from north through east, degrees in [0, 360).",
            "title": "Az Deg",
            "type": "number"
          },
          "dec_date_deg": {
            "description": "Declination of date, degrees.",
            "title": "Dec Date Deg",
            "type": "number"
          },
          "dec_icrs_deg": {
            "description": "ICRS declination, degrees.",
            "title": "Dec Icrs Deg",
            "type": "number"
          },
          "diam_deg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Apparent angular diameter, degrees (bodies only).",
            "title": "Diam Deg"
          },
          "dist_au": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Distance, au (bodies and minor bodies).",
            "title": "Dist Au"
          },
          "id": {
            "description": "Target id as requested: `hip:<n>`, `dso:<id>`, a body id, `a:<..>`, `c:<..>`.",
            "examples": [
              "hip:32349"
            ],
            "title": "Id",
            "type": "string"
          },
          "mag": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Apparent V magnitude when known.",
            "title": "Mag"
          },
          "phase": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Illuminated fraction (bodies and minor bodies).",
            "title": "Phase"
          },
          "ra_date_deg": {
            "description": "Right ascension on the true equator and equinox of date, degrees in [0, 360).",
            "title": "Ra Date Deg",
            "type": "number"
          },
          "ra_icrs_deg": {
            "description": "ICRS right ascension, degrees in [0, 360).",
            "title": "Ra Icrs Deg",
            "type": "number"
          }
        },
        "required": [
          "id",
          "alt_deg",
          "az_deg",
          "ra_icrs_deg",
          "dec_icrs_deg",
          "ra_date_deg",
          "dec_date_deg"
        ],
        "title": "AltAzEntry",
        "type": "object"
      },
...
      "MinorBodySummary": {
        "additionalProperties": false,
        "description": "One row of `/minor-bodies/search` and `/minor-bodies/defaults` (brief l.152-155).",
        "properties": {
          "designation": {
            "examples": [
              "(1) Ceres"
            ],
            "title": "Designation",
            "type": "string"
          },
          "elements_epoch_tt": {
            "description": "Epoch of the orbital elements, TT Julian Date.",
            "title": "Elements Epoch Tt",
            "type": "number"
          },
          "h_mag": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Absolute magnitude H (asteroids).",
            "title": "H Mag"
          },
          "id": {
            "description": "`a:<number>`, `a:<packed designation>` or `c:<designation>` (URL-safe).",
            "examples": [
              "a:1",
              "c:1P"
            ],
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "asteroid",
              "comet"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Ceres"
            ],
            "title": "Name"
          }
        },
        "required": [
          "id",
          "designation",
          "kind",
          "elements_epoch_tt"
        ],
        "title": "MinorBodySummary",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "RFC 9457 problem document (`type`, `title`, `status`, `detail`, `instance`, `errors[]`).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Explanation specific to this occurrence.",
            "examples": [
              "unknown body ids: vulcan"
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Per-parameter validation errors (400 only).",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Path of the request (never its query string).",
            "examples": [
              "/api/v1/sky/frame"
            ],
            "title": "Instance"
          },
          "range_tt": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Valid TT Julian Date range when the request fell outside data coverage (422).",
            "examples": [
              [
                2396758.5,
                2506000.5
              ]
            ],
            "title": "Range Tt"
          },
          "status": {
            "description": "HTTP status code.",
            "examples": [
              400
            ],
            "maximum": 599,
            "minimum": 400,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short summary of the problem type.",
            "examples": [
              "Invalid parameter"
            ],
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URI identifying the problem type; every type is documented in docs/api.md.",
            "examples": [
              "https://github.com/j-about/InterSidera/blob/master/docs/api.md#problem-invalid-parameter"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Stateless, cacheable GET API behind InterSidera: Skyfield is the astronomical authority, clients only rotate, interpolate and apply the corrections it supplies. Angles in degrees, distances in au, times as TT Julian Dates unless suffixed `_utc`. Errors are RFC 9457 problem documents (`application/problem+json`).",
    "title": "InterSidera Sky API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/minor-bodies/defaults": {
      "get": {
        "description": "At most `limits.max_minor_bodies` (100) entries in the shape of `/minor-bodies/search`: the asteroids with H <= 9 sorted by H (Pluto excluded: it is served as a major body), then the comets whose orbital-element epoch lies within 2 years of the server date, sorted by the proximity of their perihelion to the server date. Depends on the server clock only; `Cache-Control: public, max-age=300`; 503 with `Retry-After` while the MPC data is missing.",
        "operationId": "default_minor_bodies_api_v1_minor_bodies_defaults_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MinorBodySummary"
                  },
                  "title": "Response Default Minor Bodies Api V1 Minor Bodies Defaults Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Default minor bodies ranked by expected brightness",
        "tags": [
          "minor-bodies"
        ]
      }
    },
    "/api/v1/minor-bodies/search": {
      "get": {
        "description": "Case-insensitive search over the MPC orbit tables: an exact id (`a:433`, `c:1P`), a bare number (`433`), a short comet form (`1P`), a designation prefix (`2024 YR`, `C/2023`) or a substring of a name (`ceres`, `hale`). Results carry the URL-safe id to use in `/sky/frame` and `/sky/altaz`, the MPC designation, the name when the object has one, the kind, the absolute magnitude H (asteroids) and the epoch of the orbital elements. Name matches are ranked by brightness; at most `limit` entries. Rate-limited per client IP (429 with `Retry-After`); `Cache-Control: public, max-age=300`; 503 with `Retry-After` while the MPC data is missing.",
        "operationId": "search_minor_bodies_api_v1_minor_bodies_search_get",
        "parameters": [
          {
            "description": "Id, number, designation prefix or name substring; whitespace is collapsed and the match is case-insensitive.",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Id, number, designation prefix or name substring; whitespace is collapsed and the match is case-insensitive.",
              "examples": [
                "ceres",
                "433",
                "2024 YR",
                "1P",
                "C/2023 A3"
              ],
              "maxLength": 64,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Maximum number of results.",
              "examples": [
                20
              ],
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MinorBodySummary"
                  },
                  "title": "Response Search Minor Bodies Api V1 Minor Bodies Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameter (`errors[]` lists each one)."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; retry after `Retry-After` seconds."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Search asteroids and comets by id, designation or name",
        "tags": [
          "minor-bodies"
        ]
      }
    },
    "/api/v1/sky/altaz": {
      "get": {
        "description": "Skyfield's apparent altitude and azimuth, ICRS and of-date right ascension and declination for up to 200 targets seen from the observer at `tt`: `hip:<number>`, `dso:<canonical OpenNGC id>` (Messier aliases resolved through the catalog), body ids, `a:<number>` and `c:<designation>`. Bodies and minor bodies also carry `dist_au`, `mag` and `phase` (bodies: `diam_deg` too). Refraction (standard atmosphere) is applied only when requested and only for Earth observers. Rows follow the canonical (sorted) target list and echo the ids as requested. Rate limited per client IP; `Cache-Control: public, max-age=300`.",
        "operationId": "get_altaz_api_v1_sky_altaz_get",
        "parameters": [
          {
            "description": "Observer body id: `earth`, `moon`, `mercury`, `venus`, `mars`, `jupiter`, `saturn`, `uranus`, `neptune` or `pluto` (the Sun is never an observer).",
            "in": "query",
            "name": "body",
            "required": false,
            "schema": {
              "default": "earth",
              "description": "Observer body id: `earth`, `moon`, `mercury`, `venus`, `mars`, `jupiter`, `saturn`, `uranus`, `neptune` or `pluto` (the Sun is never an observer).",
              "examples": [
                "earth"
              ],
              "maxLength": 16,
              "title": "Body",
              "type": "string"
            }
          },
          {
            "description": "Observer latitude in degrees (geodetic on Earth, planetocentric elsewhere), canonicalized to 1e-6 degree.",
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "description": "Observer latitude in degrees (geodetic on Earth, planetocentric elsewhere), canonicalized to 1e-6 degree.",
              "examples": [
                48.8566
              ],
              "maximum": 90.0,
              "minimum": -90.0,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "description": "Observer east longitude in degrees within [-360, 360], canonicalized to 1e-6 degree and wrapped into [-180, 180).",
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "description": "Observer east longitude in degrees within [-360, 360], canonicalized to 1e-6 degree and wrapped into [-180, 180).",
              "examples": [
                2.3522
              ],
              "maximum": 360.0,
              "minimum": -360.0,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "description": "Observer elevation above the reference surface in metres, canonicalized to 1 m.",
            "in": "query",
            "name": "elev",
            "required": false,
            "schema": {
              "default": 0.0,
              "description": "Observer elevation above the reference surface in metres, canonicalized to 1 m.",
              "examples": [
                35
              ],
              "maximum": 100000.0,
              "minimum": -12000.0,
              "title": "Elev",
              "type": "number"
            }
          },
          {
            "description": "TT Julian Date of the instant, canonicalized to 1e-8 day.",
            "in": "query",
            "name": "tt",
            "required": true,
            "schema": {
              "description": "TT Julian Date of the instant, canonicalized to 1e-8 day.",
              "examples": [
                2461285.5
              ],
              "title": "Tt",
              "type": "number"
            }
          },
          {
            "description": "Comma-separated targets (`hip:<number>`, `dso:<id>`, body id, `a:<..>`, `c:<..>`), at most 200.",
            "in": "query",
            "name": "targets",
            "required": true,
            "schema": {
              "description": "Comma-separated targets (`hip:<number>`, `dso:<id>`, body id, `a:<..>`, `c:<..>`), at most 200.",
              "examples": [
                "hip:32349,moon,dso:NGC224,a:1,c:1P"
              ],
              "maxLength": 8192,
              "title": "Targets",
              "type": "string"
            }
          },
          {
            "description": "Apply Skyfield's standard-atmosphere refraction (Earth observers only).",
            "in": "query",
            "name": "refraction",
            "required": false,
            "schema": {
              "default": false,
              "description": "Apply Skyfield's standard-atmosphere refraction (Earth observers only).",
              "examples": [
                false
              ],
              "title": "Refraction",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AltAzEntry"
                  },
                  "title": "Response Get Altaz Api V1 Sky Altaz Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameter (`errors[]` lists each one)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unknown object."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Outside data coverage; `range_tt` carries the valid TT range."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; retry after `Retry-After` seconds."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Authoritative alt/az and RA/Dec of targets at one instant",
        "tags": [
          "sky"
        ]
      }
    },
    "/api/v1/sky/frame": {
      "get": {
        "description": "One window of `n` samples spaced `step_s` seconds apart from `tt`: the ICRF -> ENU horizon quaternions, the ICRF -> equinox-of-date quaternions, the observer's barycentric velocity, the Sun direction and, per requested body and minor body, the apparent direction, distance, magnitude, illuminated fraction and angular diameter. Parameters are canonicalized before computing and before the cache lookup (coordinates to 1e-6 degree, elevation to 1 m, `tt` to 1e-8 day, lists sorted and de-duplicated) and the canonical values are echoed. `step_s` is clamped to the smallest per-class maximum among the requested bodies (`/meta.limits.max_step_s`) and the value used is reported. `n * (bodies + minor)` may not exceed 4096. Hard coverage limits (ephemeris, observer frame) answer 422 with the valid range; soft ones are `warnings`. Rate limited per client IP; `Cache-Control: public, max-age=300`; `Server-Timing` reports the compute time and the cache outcome.",
        "operationId": "get_frame_api_v1_sky_frame_get",
        "parameters": [
          {
            "description": "Observer body id: `earth`, `moon`, `mercury`, `venus`, `mars`, `jupiter`, `saturn`, `uranus`, `neptune` or `pluto` (the Sun is never an observer).",
            "in": "query",
            "name": "body",
            "required": false,
            "schema": {
              "default": "earth",
              "description": "Observer body id: `earth`, `moon`, `mercury`, `venus`, `mars`, `jupiter`, `saturn`, `uranus`, `neptune` or `pluto` (the Sun is never an observer).",
              "examples": [
                "earth"
              ],
              "maxLength": 16,
              "title": "Body",
              "type": "string"
            }
          },
          {
            "description": "Observer latitude in degrees (geodetic on Earth, planetocentric elsewhere), canonicalized to 1e-6 degree.",
            "in": "query",
            "name": "lat",
            "required": true,
            "schema": {
              "description": "Observer latitude in degrees (geodetic on Earth, planetocentric elsewhere), canonicalized to 1e-6 degree.",
              "examples": [
                48.8566
              ],
              "maximum": 90.0,
              "minimum": -90.0,
              "title": "Lat",
              "type": "number"
            }
          },
          {
            "description": "Observer east longitude in degrees within [-360, 360], canonicalized to 1e-6 degree and wrapped into [-180, 180).",
            "in": "query",
            "name": "lon",
            "required": true,
            "schema": {
              "description": "Observer east longitude in degrees within [-360, 360], canonicalized to 1e-6 degree and wrapped into [-180, 180).",
              "examples": [
                2.3522
              ],
              "maximum": 360.0,
              "minimum": -360.0,
              "title": "Lon",
              "type": "number"
            }
          },
          {
            "description": "Observer elevation above the reference surface in metres, canonicalized to 1 m.",
            "in": "query",
            "name": "elev",
            "required": false,
            "schema": {
              "default": 0.0,
              "description": "Observer elevation above the reference surface in metres, canonicalized to 1 m.",
              "examples": [
                35
              ],
              "maximum": 100000.0,
              "minimum": -12000.0,
              "title": "Elev",
              "type": "number"
            }
          },
          {
            "description": "TT Julian Date of the first sample, canonicalized to 1e-8 day; any finite value is accepted and the data coverage decides (422 outside it).",
            "in": "query",
            "name": "tt",
            "required": true,
            "schema": {
              "description": "TT Julian Date of the first sample, canonicalized to 1e-8 day; any finite value is accepted and the data coverage decides (422 outside it).",
              "examples": [
                2461285.5
              ],
              "title": "Tt",
              "type": "number"
            }
          },
          {
            "description": "Seconds between samples; clamped to the requested bodies' class maximum and echoed in `time.step_s`.",
            "in": "query",
            "name": "step_s",
            "required": false,
            "schema": {
              "default": 60,
              "description": "Seconds between samples; clamped to the requested bodies' class maximum and echoed in `time.step_s`.",
              "examples": [
                300
              ],
              "maximum": 31557600,
              "minimum": 1,
              "title": "Step S",
              "type": "integer"
            }
          },
          {
            "description": "Number of samples.",
            "in": "query",
            "name": "n",
            "required": false,
            "schema": {
              "default": 32,
              "description": "Number of samples.",
              "examples": [
                32
              ],
              "maximum": 64,
              "minimum": 1,
              "title": "N",
              "type": "integer"
            }
          },
          {
            "description": "`all` (every body but the observer's own) or a comma-separated list of body ids; naming the observer's body is a 400.",
            "in": "query",
            "name": "bodies",
            "required": false,
            "schema": {
              "default": "all",
              "description": "`all` (every body but the observer's own) or a comma-separated list of body ids; naming the observer's body is a 400.",
              "examples": [
                "all",
                "moon,mars,jupiter"
              ],
              "maxLength": 256,
              "title": "Bodies",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated minor-body ids (`a:<number>`, `a:<packed designation>`, `c:<designation>`), at most 100.",
            "in": "query",
            "name": "minor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated minor-body ids (`a:<number>`, `a:<packed designation>`, `c:<designation>`), at most 100.",
              "examples": [
                "a:1,c:1P"
              ],
              "title": "Minor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameter (`errors[]` lists each one)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unknown object."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Outside data coverage; `range_tt` carries the valid TT range."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Rate limited; retry after `Retry-After` seconds."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Data not ready (starting, or degraded for this endpoint); retry after `Retry-After`."
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Any other error, as an RFC 9457 problem document."
          }
        },
        "summary": "Sky frame: rotations and body samples over a time window",
        "tags": [
          "sky"
        ]
      }
    }
  },
...
}